refactor: split dce if pruning
